pyproject.toml
[project]
name = "flask-rest-api-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "pyjwt>=2.10.1",
    "pytest>=9.0.2",
    "werkzeug>=3.1.5",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["test"]

